+++ build_runner/test/build_plan/build_phases_test.dart
@@ -13,28 +13,37 @@ import '../common/common.dart';
 void main() {
   group('BuildPhases', () {
     test('digest is equal for equal phases', () {
-      final buildPhases1 = BuildPhases([InBuildPhase(TestBuilder(), 'a')]);
-      final buildPhases2 = BuildPhases([InBuildPhase(TestBuilder(), 'a')]);
+      final buildPhases1 = BuildPhases([
+        InBuildPhase(builder: TestBuilder(), key: 'TestBuilder', package: 'a'),
+      ]);
+      final buildPhases2 = BuildPhases([
+        InBuildPhase(builder: TestBuilder(), key: 'TestBuilder', package: 'a'),
+      ]);
 
       expect(buildPhases1.digest, buildPhases2.digest);
     });
 
     test('digest changes on additional builder', () {
-      final buildPhases1 = BuildPhases([InBuildPhase(TestBuilder(), 'a')]);
+      final buildPhases1 = BuildPhases([
+        InBuildPhase(builder: TestBuilder(), key: 'TestBuilder', package: 'a'),
+      ]);
       final buildPhases2 = BuildPhases([
-        InBuildPhase(TestBuilder(), 'a'),
-        InBuildPhase(TestBuilder(), 'a'),
+        InBuildPhase(builder: TestBuilder(), key: 'TestBuilder', package: 'a'),
+        InBuildPhase(builder: TestBuilder(), key: 'TestBuilder', package: 'a'),
       ]);
 
       expect(buildPhases1.digest, isNot(buildPhases2.digest));
     });
 
     test('digest changes on extension change', () {
-      final buildPhases1 = BuildPhases([InBuildPhase(TestBuilder(), 'a')]);
+      final buildPhases1 = BuildPhases([
+        InBuildPhase(builder: TestBuilder(), key: 'TestBuilder', package: 'a'),
+      ]);
       final buildPhases2 = BuildPhases([
         InBuildPhase(
-          TestBuilder(buildExtensions: appendExtension('different')),
-          'a',
+          builder: TestBuilder(buildExtensions: appendExtension('different')),
+          key: 'TestBuilder',
+          package: 'a',
         ),
       ]);
 
@@ -44,8 +53,12 @@ void main() {
     test('digest does not change on builder change', () {
       // Changes to builder code is checked via changes to the build script and
       // deps, not by `BuildPhases`.
-      final buildPhases1 = BuildPhases([InBuildPhase(TestBuilder(), 'a')]);
-      final buildPhases2 = BuildPhases([InBuildPhase(TestBuilder2(), 'a')]);
+      final buildPhases1 = BuildPhases([
+        InBuildPhase(builder: TestBuilder(), key: 'TestBuilder', package: 'a'),
+      ]);
+      final buildPhases2 = BuildPhases([
+        InBuildPhase(builder: TestBuilder2(), key: 'TestBuilder', package: 'a'),
+      ]);
 
       expect(buildPhases1.digest, buildPhases2.digest);
     });
@@ -53,12 +66,15 @@ void main() {
     test('digest does not change on builder options change', () {
       // Changes to builder code is checked via changes to the build script and
       // deps, not by `BuildPhases`.
-      final buildPhases1 = BuildPhases([InBuildPhase(TestBuilder(), 'a')]);
+      final buildPhases1 = BuildPhases([
+        InBuildPhase(builder: TestBuilder(), key: 'TestBuilder', package: 'a'),
+      ]);
       final buildPhases2 = BuildPhases([
         InBuildPhase(
-          TestBuilder(),
-          'a',
-          builderOptions: const BuilderOptions({'a': 'b'}),
+          builder: TestBuilder(),
+          key: 'TestBuilder',
+          package: 'a',
+          options: const BuilderOptions({'a': 'b'}),
         ),
       ]);
 
@@ -68,12 +84,15 @@ void main() {
     test('options digest changes on builder options change', () {
       // Changes to builder code is checked via changes to the build script and
       // deps, not by `BuildPhases`.
-      final buildPhases1 = BuildPhases([InBuildPhase(TestBuilder(), 'a')]);
+      final buildPhases1 = BuildPhases([
+        InBuildPhase(builder: TestBuilder(), key: 'TestBuilder', package: 'a'),
+      ]);
       final buildPhases2 = BuildPhases([
         InBuildPhase(
-          TestBuilder(),
-          'a',
-          builderOptions: const BuilderOptions({'a': 'b'}),
+          builder: TestBuilder(),
+          key: 'TestBuilder',
+          package: 'a',
+          options: const BuilderOptions({'a': 'b'}),
         ),
       ]);
 
@@ -90,9 +109,9 @@ void main() {
         [],
         PostBuildPhase([
           PostBuildAction(
-            const FileDeletingBuilder(['']),
-            'a',
-            builderOptions: const BuilderOptions({}),
+            builder: const FileDeletingBuilder(['']),
+            package: 'a',
+            options: const BuilderOptions({}),
             targetSources: const InputSet(),
             generateFor: const InputSet(),
           ),
@@ -102,9 +121,9 @@ void main() {
         [],
         PostBuildPhase([
           PostBuildAction(
-            const FileDeletingBuilder(['']),
-            'a',
-            builderOptions: const BuilderOptions({'a': 'b'}),
+            builder: const FileDeletingBuilder(['']),
+            package: 'a',
+            options: const BuilderOptions({'a': 'b'}),
             targetSources: const InputSet(),
             generateFor: const InputSet(),
           ),